+++ components/src/lib/Modal/index.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from "react";
-import useOutsideClickRef from "../../hooks/useOutsideClickRef";
+import useOutsideClickRef from "@hooks/useOutsideClickRef";
 import useKeyDown from "@/hooks/useKeyDown";
 import BaseModal from "../BaseModal";
 import { ModalLayoutProps, ModalDefaultProps } from "../types";
