+++ hooks/jest.config.cjs
@@ -1,4 +1,16 @@
 module.exports = {
   preset: "ts-jest",
   testEnvironment: "jsdom",
+  globals: {
+    "ts-jest": {
+      tsconfig: "tsconfig.app.json",
+    },
+  },
+  moduleNameMapper: {
+    "^@/(.*)$": "<rootDir>/src/$1",
+    "^@hooks/(.*)$": "<rootDir>/src/hooks/$1",
+    "^@tests/(.*)$": "<rootDir>/src/tests/$1",
+    "^@utils/(.*)$": "<rootDir>/src/utils/$1",
+    "^@lib/(.*)$": "<rootDir>/src/lib/$1",
+  },
 };
