+++ hooks/src/lib/hooks/CardExpiryDate/utils/index.ts
@@ -1,8 +1,5 @@
-import {
-  commonConditions,
-  getErrorByRules,
-} from "../../../../utils/validation";
-import { ErrorState, ValidationRule } from "../../../types";
+import { commonConditions, getErrorByRules } from "@utils/validation";
+import { ErrorState, ValidationRule } from "@lib/types";
 
 const cardExpiryDateConditions = {
   isValidMonth: (value: string) => {
