+++ hooks/src/lib/hooks/CardPassword/utils/index.ts
@@ -1,8 +1,5 @@
-import { ErrorState, ValidationRule } from "../../../types";
-import {
-  commonConditions,
-  getErrorByRules,
-} from "../../../../utils/validation";
+import { ErrorState, ValidationRule } from "@lib/types";
+import { commonConditions, getErrorByRules } from "@utils/validation";
 
 const cardPasswordValidationRules: ValidationRule<string>[] = [
   {
